Move lombscargle statistics into test_lombscargle
