+++ astropy/stats/lombscargle/tests/test_lombscargle.py
@@ -325,3 +325,196 @@ def test_autopower(data):
 
     assert_allclose(freq1, freq2)
     assert_allclose(power1, power2)
+
+
+@pytest.fixture
+def null_data(N=1000, dy=1, rseed=0):
+    """Generate null hypothesis data"""
+    rng = np.random.RandomState(rseed)
+    t = 100 * rng.rand(N)
+    dy = 0.5 * dy * (1 + rng.rand(N))
+    y = dy * rng.randn(N)
+    return t, y, dy
+
+
+@pytest.mark.parametrize('normalization', NORMALIZATIONS)
+def test_distribution(null_data, normalization):
+    t, y, dy = null_data
+    N = len(t)
+    ls = LombScargle(t, y, dy)
+    freq, power = ls.autopower(normalization=normalization,
+                               maximum_frequency=40)
+    z = np.linspace(0, power.max(), 1000)
+
+    # Test that pdf and cdf are consistent
+    dz = z[1] - z[0]
+    z_mid = z[:-1] + 0.5 * dz
+    pdf = _lombscargle_pdf(z_mid, N, normalization=normalization)
+    cdf = _lombscargle_cdf(z, N, normalization=normalization)
+    assert_allclose(pdf, np.diff(cdf) / dz, rtol=1E-5, atol=1E-8)
+
+    # Test that observed power is distributed according to the theoretical pdf
+    hist, bins = np.histogram(power, 30, normed=True)
+    midpoints = 0.5 * (bins[1:] + bins[:-1])
+    pdf = _lombscargle_pdf(midpoints, N, normalization=normalization)
+    assert_allclose(hist, pdf, rtol=0.05, atol=0.05 * pdf[0])
+
+
+def test_distribution_units(null_data):
+    t, y, dy = null_data
+    N = len(t)
+
+    t_days = t * units.day
+    y_mag = y * units.mag
+    dy_mag = dy * units.mag
+
+    # unnormalized: this should fail
+    ls = LombScargle(t_days, y_mag)
+    freq, power = ls.autopower(normalization='psd',
+                               maximum_frequency=1 / units.day)
+    assert power.unit == units.mag ** 2
+
+    with pytest.raises(ValueError) as err:
+        pdf = _lombscargle_pdf(power, N, normalization='psd')
+    assert str(err.value).startswith('The distribution can be computed')
+
+    with pytest.raises(ValueError) as err:
+        cdf = _lombscargle_cdf(power, N, normalization='psd')
+    assert str(err.value).startswith('The distribution can be computed')
+
+    # normalized: case with units should match case without units
+    ls = LombScargle(t_days, y_mag, dy_mag)
+    freq, power = ls.autopower(normalization='psd',
+                               maximum_frequency=1 / units.day)
+    assert power.unit == units.dimensionless_unscaled
+
+    pdf_with_units = _lombscargle_pdf(power, N, normalization='psd')
+    cdf_with_units = _lombscargle_cdf(power, N, normalization='psd')
+
+    ls = LombScargle(t, y, dy)
+    freq, power = ls.autopower(normalization='psd',
+                               maximum_frequency=1)
+    pdf_no_units = _lombscargle_pdf(power, N, normalization='psd')
+    cdf_no_units = _lombscargle_cdf(power, N, normalization='psd')
+
+    assert_allclose(pdf_with_units, pdf_no_units)
+    assert_allclose(cdf_with_units, cdf_no_units)
+
+
+# The following are convenience functions used to compute statistics of the
+# periodogram under various normalizations; they are used in the preceding
+# two tests.
+
+
+def _lombscargle_pdf(z, N, normalization, dH=1, dK=3):
+    """Probability density function for Lomb-Scargle periodogram
+
+    Compute the expected probability density function of the periodogram
+    for the null hypothesis - i.e. data consisting of Gaussian noise.
+
+    Parameters
+    ----------
+    z : array-like
+        the periodogram value
+    N : int
+        the number of data points from which the periodogram was computed
+    normalization : string
+        The periodogram normalization. Must be one of
+        ['standard', 'model', 'log', 'psd']
+    dH, dK : integers (optional)
+        The number of parameters in the null hypothesis and the model
+
+    Returns
+    -------
+    pdf : np.ndarray
+        The expected probability density function
+
+    Notes
+    -----
+    For normalization='psd', the distribution can only be computed for
+    periodograms constructed with errors specified.
+    All expressions used here are adapted from Table 1 of Baluev 2008 [1]_.
+
+    References
+    ----------
+    .. [1] Baluev, R.V. MNRAS 385, 1279 (2008)
+    """
+    if dK - dH != 2:
+        raise NotImplementedError("Degrees of freedom != 2")
+    Nk = N - dK
+
+    if isinstance(z, units.Quantity):
+        if z.unit == units.dimensionless_unscaled:
+            z = z.value
+        else:
+            raise ValueError('The distribution can be computed only for '
+                             'normalized (unitless) periodograms.')
+
+    if normalization == 'psd':
+        return np.exp(-z)
+    elif normalization == 'standard':
+        return 0.5 * Nk * (1 + z) ** (-0.5 * Nk - 1)
+    elif normalization == 'model':
+        return 0.5 * Nk * (1 - z) ** (0.5 * Nk - 1)
+    elif normalization == 'log':
+        return 0.5 * Nk * np.exp(-0.5 * Nk * z)
+    else:
+        raise ValueError("normalization='{0}' is not recognized"
+                         "".format(normalization))
+
+
+def _lombscargle_cdf(z, N, normalization, dH=1, dK=3):
+    """Cumulative distribution for the Lomb-Scargle periodogram
+
+    Compute the expected cumulative distribution of the periodogram
+    for the null hypothesis - i.e. data consisting of Gaussian noise.
+
+    Parameters
+    ----------
+    z : array-like
+        the periodogram value
+    N : int
+        the number of data points from which the periodogram was computed
+    normalization : string
+        The periodogram normalization. Must be one of
+        ['standard', 'model', 'log', 'psd']
+    dH, dK : integers (optional)
+        The number of parameters in the null hypothesis and the model
+
+    Returns
+    -------
+    cdf : np.ndarray
+        The expected cumulative distribution function
+
+    Notes
+    -----
+    For normalization='psd', the distribution can only be computed for
+    periodograms constructed with errors specified.
+    All expressions used here are adapted from Table 1 of Baluev 2008 [1]_.
+
+    References
+    ----------
+    .. [1] Baluev, R.V. MNRAS 385, 1279 (2008)
+    """
+    if dK - dH != 2:
+        raise NotImplementedError("Degrees of freedom != 2")
+    Nk = N - dK
+
+    if isinstance(z, units.Quantity):
+        if z.unit == units.dimensionless_unscaled:
+            z = z.value
+        else:
+            raise ValueError('The distribution can be computed only for '
+                             'normalized (unitless) periodograms.')
+
+    if normalization == 'psd':
+        return 1 - np.exp(-z)
+    elif normalization == 'standard':
+        return 1 - (1 + z) ** (-0.5 * Nk)
+    elif normalization == 'model':
+        return 1 - (1 - z) ** (0.5 * Nk)
+    elif normalization == 'log':
+        return 1 - np.exp(-0.5 * Nk * z)
+    else:
+        raise ValueError("normalization='{0}' is not recognized"
+                         "".format(normalization))
